refactor(realitea): split repository.server.ts into domain modules
Splits the 812-line data-access monolith into six focused modules
(games, puzzles, articles, generation-runs, admin-actions, attempts)
under app/lib/realitea/server/, updates all 26 call sites and test
mocks, and adds per-module test coverage for the ~20 exports that
previously had none. getPendingArticlesForGame now delegates to
getPendingArticlesForTopics instead of duplicating the query.

No behavior changes.
