+++ app/lib/realitea/server/admin-actions.server.ts
@@ -1,0 +1,44 @@
+/**
+ * Data access for `realitea_admin_actions` — the audit trail of operator
+ * writes (generate, publish, replace, etc) made through the admin console.
+ */
+
+import type { RealiteaAdminActionKind } from "~/lib/server/db";
+import { adminActions, and, count, db, eq, gte, inArray } from "~/lib/server/db";
+
+export async function countRecentGenerateActions(
+  hominemUserId: string,
+  since: Date,
+): Promise<number> {
+  const [row] = await db
+    .select({ value: count() })
+    .from(adminActions)
+    .where(
+      and(
+        eq(adminActions.hominemUserId, hominemUserId),
+        inArray(adminActions.kind, ["generate", "preview"]),
+        gte(adminActions.at, since),
+      ),
+    );
+  return row?.value ?? 0;
+}
+
+export async function recordAdminAction(input: {
+  hominemUserId?: string;
+  kind: RealiteaAdminActionKind;
+  gamesTopicId: number;
+  dateUtc?: string;
+  dryRun?: boolean;
+  payload?: Record<string, unknown>;
+  result?: Record<string, unknown>;
+}): Promise<void> {
+  await db.insert(adminActions).values({
+    hominemUserId: input.hominemUserId ?? "system:generate",
+    kind: input.kind,
+    gamesTopicId: input.gamesTopicId,
+    dateUtc: input.dateUtc,
+    dryRun: input.dryRun ?? false,
+    payload: input.payload ?? {},
+    result: input.result ?? {},
+  });
+}
