+++ app/lib/realitea/server/history.server.ts
@@ -1,12 +1,8 @@
 import type { GameStatus, RealiteaGuess, PuzzleAnswerType } from "../core/types";
 import { addDaysToDateKey, daysBetweenDateKeys, getDateKey } from "../core/date";
-import {
-  getEarliestPuzzleDateKey,
-  getExistingDateKeys,
-  getGameBySlug,
-  listAttemptsForUserInRange,
-  loadAllAttemptsForUser,
-} from "./repository.server";
+import { listAttemptsForUserInRange, loadAllAttemptsForUser } from "./attempts.server";
+import { getGameBySlug } from "./games.server";
+import { getEarliestPuzzleDateKey, getExistingDateKeys } from "./puzzles.server";
 import { computeHistoryStats, type PuzzleHistoryStats } from "../core/stats";
 import { DEFAULT_REALITEA_GAME_SLUG } from "../generation/catalog";
