+++ scripts/realitea-e2e.ts
@@ -7,11 +7,8 @@ import { chromium, type Page } from "playwright";
 
 import { getDateKey } from "../app/lib/realitea/core/date";
 import { DEFAULT_REALITEA_GAME_SLUG } from "../app/lib/realitea/generation/catalog";
-import {
-  getGameBySlug,
-  loadMostRecentPuzzle,
-  loadPuzzleForDate,
-} from "../app/lib/realitea/server/repository.server";
+import { getGameBySlug } from "../app/lib/realitea/server/games.server";
+import { loadMostRecentPuzzle, loadPuzzleForDate } from "../app/lib/realitea/server/puzzles.server";
 import { runScript } from "./_shared/run-script";
 
 const DEFAULT_BASE_URL = "http://localhost:3001";
